+++ src/app/api/users/[uid]/route.ts
@@ -1,37 +1,10 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { feuerwehren } from '../../../../common/feuerwehren';
 import { UserRecordExtended } from '../../../../common/users';
 import adminRequired from '../../../../server/auth/adminRequired';
-import { firestore } from '../../../../server/firebase/admin';
+import { updateUser } from './updateUser';
 
 export interface UsersResponse {
   user: UserRecordExtended;
-}
-
-export async function updateUser(uid: string, user: UserRecordExtended) {
-  const newData = {
-    displayName: user.displayName,
-    email: user.email,
-    authorized: user.authorized,
-    feuerwehr: user.feuerwehr || 'neusiedl',
-    abschnitt: feuerwehren[user.feuerwehr || 'neusiedl'].abschnitt || 0,
-  };
-
-  console.info(`updating ${uid}: ${JSON.stringify(newData)}`);
-
-  await firestore
-    .collection('user')
-    .doc(`${uid}`)
-    .set(
-      Object.fromEntries(
-        Object.entries(newData).filter(([key, value]) => key && value)
-      ),
-      {
-        merge: true,
-      }
-    );
-
-  return { ...newData, ...user };
 }
 
 export async function POST(
